fix(parser): allow Long values for tempo attribute
In our dataset, there is one project that uses a larger-than-Integer
value.

+++ src/main/java/de/uni_passau/fim/se2/litterbox/ast/new_parser/raw_ast/RawTarget.java
@@ -217,7 +217,7 @@ public final class RawTarget {
     }
 
     public record StageAttributes(
-            Integer tempo, String videoState, Double videoTransparency, String textToSpeechLanguage
+            Long tempo, String videoState, Double videoTransparency, String textToSpeechLanguage
     ) {
     }
